% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grademebsc.R
\name{bsc_grademe}
\alias{bsc_grademe}
\title{R function to grade marks}
\usage{
bsc_grademe(marks)
}
\arguments{
\item{marks}{of marks}
}
\value{
vector containing grades
}
\description{
R function to grade marks
}
