% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mid_end.R
\name{mid_end}
\alias{mid_end}
\title{R function to grade marks}
\usage{
mid_end(mid, end, var)
}
\arguments{
\item{mid}{marks of mid and assignments}

\item{end}{marks of end assignment}

\item{var}{a character vector of variables to join by}
}
\value{
table combining marks of mid and end
}
\description{
R function to grade marks
}
