% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grademe.R
\name{pg_grademe}
\alias{pg_grademe}
\title{R function to grade marks}
\usage{
pg_grademe(marks)
}
\arguments{
\item{marks}{of marks}
}
\value{
vector containing grades
}
\description{
R function to grade marks
}
