% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readmarks.R
\name{readmarks}
\alias{readmarks}
\title{R function to read marks table in the excel sheet}
\usage{
readmarks(path, sheetIndex, colrange, rwrange)
}
\arguments{
\item{path}{path to xlsx file}

\item{sheetIndex}{number representing the sheet index in the workbook}

\item{colrange}{column range}

\item{rwrange}{raw range}
}
\value{
table of marks information
}
\description{
R function to read marks table in the excel sheet
}
