% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readmarks.R
\name{readmarks}
\alias{readmarks}
\title{R function to read marks table in the excel sheet}
\usage{
readmarks(path, sheet, range)
}
\arguments{
\item{path}{path to xlsx file}

\item{sheet}{Sheet to read. Either a string (the name of a sheet), or an integer (the position of the sheet}

\item{range}{column range string for example "R6C1:R29C2". This reads raws from R6 to R29 and columns from C1 to C2.}
}
\value{
table of marks information
}
\description{
R function to read marks table in the excel sheet
}
